pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo_rag"
version = "0.0.1"
description = "Codebase for Repo RAG"
requires-python = ">=3.8,<3.13"
dependencies = [
    "jupyter~=1.0.0",
    "jupyterlab~=3.6",
    "langchain==0.3.9",   
    "transformers>=4.0.0",
    "sentence-transformers>=3.4.0",
    "torch>=2.0.0",         
    "langchain-community==0.3.8", 
    "langchain-openai==0.2.10",   
    "python-dotenv~=1.0.1",
    "faiss-cpu~=1.8.0.post1",
    "pydantic~=2.8.2",
    "langchain-core==0.3.21", 
    "langgraph~=0.2.53",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "requests",
    "streamlit",
]



[project.urls]
"Source" = "https://github.com/Micz26/RepoRAG"

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    # "UP",  # pyupgrade
]
line-length = 120
ignore = [
    # "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]
extend-exclude = [".venv/", "build/", "dist/", "docs/"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint.isort]
order-by-type = true
lines-between-types = 1
combine-as-imports = true
# known_first_party = "common"

[tool.setuptools.dynamic]
readme = {file = ['README.md']}
